BG2-2526: Move CorsMiddleware into Middleware namesapce
